Resolved:     - Add endpoint that return http 200
    - Fix some titles

    - there is more sub event in the pipeline event, add them

+++ src/main/java/dev/idan/bgbot/data/pipeline/PipelineBuilds.java
@@ -1,13 +1,12 @@
 package dev.idan.bgbot.data.pipeline;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import lombok.Getter;
 
 @Getter
 public class PipelineBuilds {
 
-    int id;
+    long id;
 
     String stage;
